Huge update

+++ commands/pause.js
@@ -1,18 +1,14 @@
-const { UpdateEmbed } = require('../functions/updateEmbed.js');
+const { UpdateEmbed } = require("../functions/updateEmbed.js");
+const { getGuildPlayer } = require("../functions/getGuildPlayer");
 
 module.exports = {
-	name: 'pause',
-        description: 'Pauses current song',
-        aliases: ['зфгыу'],
+  name: "pause",
+  description: "Pauses current song",
+  aliases: ["зфгыу"],
 
-	execute(message, args) {
-        if(dispatcher)
-        {
-            dispatcher.pause();
-            UpdateEmbed()
-
-            if(message.content)
-                message.delete();
-        }
-	},
+  execute(message, args) {
+    const player = getGuildPlayer(message);
+    player.pause();
+    if (message.content) message.delete();
+  },
 };
