+++ commands/resume.js
@@ -1,18 +1,18 @@
 const { UpdateEmbed } = require("../functions/updateEmbed");
 
 module.exports = {
-	name: 'resume',
-        description: 'Resumes current song',
-        alliases: ['куыгьу'],
+  name: "resume",
+  description: "Resumes current song",
+  alliases: ["куыгьу"],
 
-	execute(message, args) {
-        if(dispatcher)
-        {
-            dispatcher.resume();
-            UpdateEmbed()
+  execute(message, args) {
+    console.log("resume exec");
+    //if (dispatcher) {
+    console.log("resume dispatcher");
+    dispatcher.resume();
+    UpdateEmbed();
 
-            if(message.content)
-                message.delete();
-        }
-	},
+    if (message.content) message.delete();
+    //}
+  },
 };
